Hi,

I replaced the `while`-loop with a binary search so that the `.at()`-function is faster on larger datasets.
And I added a test to compare the speed:

```
before:

elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            0ms
100         |        1000 |            8ms |           11ms
1000        |       10000 |          909ms |          847ms
2000        |       20000 |         7275ms |         3266ms
4000        |       40000 |        56646ms |        14046ms

after:

elements in | times .at() | initialization | execution time
dataset     | is used     | time of spline | of all .at()'s
------------|-------------|----------------|---------------
10          |         100 |            0ms |            1ms
100         |        1000 |           13ms |            5ms
1000        |       10000 |          913ms |           22ms
2000        |       20000 |         7142ms |           53ms
4000        |       40000 |        56850ms |           89ms
```

The binary search itself is inspired by [an similar implementation on Stackoverflow](https://stackoverflow.com/a/40850313/4417327).