@morganherlocker 2019-12-10
Excellent! Thanks for taking the time to make this a super clean implementation. I will roll this out as a patch release.